requirements.txt
flask
yfinance
pandas
numpy
psycopg2-binary
